#28 [feat] fab 배경 가리는 배경 처리 완료
