+++ app/src/main/java/com/spark/android/ui/MainActivity.kt
@@ -10,6 +10,7 @@ import com.spark.android.ui.base.BaseActivity
 import com.spark.android.ui.home.FeedFragment
 import com.spark.android.ui.home.HomeMainFragment
 import com.spark.android.ui.home.StorageFragment
+import com.spark.android.util.FloatingAnimationUtil
 import com.spark.android.util.initStatusBarColor
 import com.spark.android.util.initStatusBarTextColorToWhite
 
@@ -18,6 +19,7 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
     private lateinit var feedFragment: FeedFragment
     private lateinit var homeMainFragment: HomeMainFragment
     private lateinit var storageFragment: StorageFragment
+    private var fabState = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -27,7 +29,8 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
         initBottomNavigationFeedButton()
         initBottomNavigationHomeButton()
         initBottomNavigationStorageButton()
-
+        initBindingVariable()
+        initFloatingButtonClickListener()
     }
 
     private fun initStatusBarStyle() {
@@ -52,6 +55,9 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
             binding.tvMainBottomNavHome.setTextColor(ContextCompat.getColor(SparkApplication.ApplicationContext(), R.color.spark_dark_gray))
             binding.btnMainBottomNavStorage.setImageResource(R.drawable.ic_bottom_navigation_storage)
             binding.tvMainBottomNavStorage.setTextColor(ContextCompat.getColor(SparkApplication.ApplicationContext(), R.color.spark_dark_gray))
+            binding.fabHomeMain.visibility = View.GONE
+            binding.fabHomeMakeRoom.visibility = View.GONE
+            binding.fabHomeJoinCode.visibility = View.GONE
         }
     }
 
@@ -64,6 +70,9 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
             binding.tvMainBottomNavHome.setTextColor(ContextCompat.getColor(SparkApplication.ApplicationContext(), R.color.spark_dark_pinkred))
             binding.btnMainBottomNavStorage.setImageResource(R.drawable.ic_bottom_navigation_storage)
             binding.tvMainBottomNavStorage.setTextColor(ContextCompat.getColor(SparkApplication.ApplicationContext(), R.color.spark_dark_gray))
+            binding.fabHomeMain.visibility = View.VISIBLE
+            binding.fabHomeMakeRoom.visibility = View.VISIBLE
+            binding.fabHomeJoinCode.visibility = View.VISIBLE
         }
     }
 
@@ -76,7 +85,22 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
             binding.tvMainBottomNavHome.setTextColor(ContextCompat.getColor(SparkApplication.ApplicationContext(), R.color.spark_dark_gray))
             binding.btnMainBottomNavStorage.setImageResource(R.drawable.ic_bottom_navigation_storage_selected)
             binding.tvMainBottomNavStorage.setTextColor(ContextCompat.getColor(SparkApplication.ApplicationContext(), R.color.spark_dark_pinkred))
+            binding.fabHomeMain.visibility = View.GONE
+            binding.fabHomeMakeRoom.visibility = View.GONE
+            binding.fabHomeJoinCode.visibility = View.GONE
         }
+    }
+
+    private fun initFloatingButtonClickListener() {
+        binding.fabHomeMain.setOnClickListener(){
+            FloatingAnimationUtil.toggleFab(binding.fabHomeMain,binding.fabHomeMakeRoom,binding.fabHomeJoinCode,binding.layoutMainFabBackground,fabState)
+            fabState = !fabState
+            initBindingVariable()
+        }
+    }
+
+    private fun initBindingVariable(){
+        binding.fabState = fabState
     }
 
 }
