+++ app/src/main/res/layout/fragment_home_main.xml
@@ -5,9 +5,7 @@
 
     <data>
 
-        <variable
-            name="fabState"
-            type="Boolean" />
+
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -66,47 +64,6 @@
             tools:layout_editor_absoluteX="0dp"
             tools:listitem="@layout/item_home_recyclerview" />
 
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab_home_join_code"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:backgroundTint="@color/spark_dark_pinkred"
-            android:src="@drawable/ic_fab_join_code"
-            app:borderWidth="0dp"
-            app:fabSize="mini"
-            app:layout_constraintBottom_toBottomOf="@+id/fab_home_main"
-            app:layout_constraintEnd_toEndOf="@+id/fab_home_main"
-            app:layout_constraintStart_toStartOf="@+id/fab_home_main"
-            app:layout_constraintTop_toTopOf="@+id/fab_home_main" />
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab_home_make_room"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:backgroundTint="@color/spark_dark_pinkred"
-            android:src="@drawable/ic_fab_make_room"
-            app:borderWidth="0dp"
-            app:fabSize="mini"
-            app:layout_constraintBottom_toBottomOf="@+id/fab_home_main"
-            app:layout_constraintEnd_toEndOf="@+id/fab_home_main"
-            app:layout_constraintStart_toStartOf="@+id/fab_home_main"
-            app:layout_constraintTop_toTopOf="@+id/fab_home_main" />
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab_home_main"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="28dp"
-            android:layout_marginBottom="28dp"
-            android:backgroundTint="@{fabState ? @color/spark_white : @color/spark_dark_pinkred}"
-            android:src="@drawable/ic_fab_plus"
-            app:borderWidth="0dp"
-            app:fabCustomSize="60dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            tools:backgroundTint="@color/spark_dark_pinkred" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
